Fixed issue where Iso HTML & State would be out-of-sync
This took me a while to figure out! I hope the fix makes sense.

It should address a problem where the HTML and state that are passed to Iso would not match up, as the two are collected at ever so slightly different points in time.

I had a component with an `inProgress` boolean as part of its state. This would of course change depending on when `onFetchStart` and `onFetchSuccess`/`onFetchFail` actions are dispatched. I could not understand why the Iso HTML markup reflected that `inProgress` was `false`, while the Iso state object reflected that `inProgress` was `true`. As it turns out it was due to the way `alt-renderer` uses `flux.flush()` *after* the second `React.renderToString`.

My thinking is that because all XHR promises will have been resolved on the first render, a snapshot of the Alt state can be taken after all promises have been resolved, and immediately before running the second `React.renderToString`.